requirements.txt
langchain>=0.2.10
langchain-community>=0.2.10
langchain-openai>=0.1.10
python-dotenv
sqlalchemy
oracledb
